server/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
numpy
pandas
umap-learn
biopython
matplotlib
seaborn
websockets
aiohttp